requirements.txt
elevenlabs>=2.39.0
python-dotenv
